backend-api/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.28.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0